% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellbender.r
\name{export_seurat_to_h5}
\alias{export_seurat_to_h5}
\title{Helper function: Export Seurat object to h5 format for CellBender}
\usage{
export_seurat_to_h5(seurat_obj, output_path)
}
\arguments{
\item{seurat_obj}{Seurat object containing raw count data}

\item{output_path}{Path for the output h5 file}
}
\value{
Path to the created h5 file
}
\description{
Helper function: Export Seurat object to h5 format for CellBender
}
