% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadObject.r
\name{ReadObject}
\alias{ReadObject}
\title{Read an RDS object from the package’s data path}
\usage{
ReadObject(name)
}
\arguments{
\item{name}{The base name (without .rds) of the file}
}
\value{
The R object read from the file
}
\description{
Read an RDS object from the package’s data path
}
